requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
decorator==5.1.1
Django==5.1.5
djangorestframework==3.15.2
drf-spectacular==0.28.0
executing==2.1.0
inflection==0.5.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.1.1
wcwidth==0.2.13
